@@ -12,45 +12,34 @@
     "start": "ng serve",
     "build": "ng build",
     "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e"
+    "lint": "ng lint"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.5",
-    "@angular/common": "^5.2.5",
-    "@angular/compiler": "^5.2.5",
-    "@angular/core": "^5.2.5",
-    "@angular/forms": "^5.2.5",
-    "@angular/platform-browser": "^5.2.5",
-    "@angular/platform-browser-dynamic": "^5.2.5",
+    "@angular/animations": "v6-lts",
+    "@angular/common": "v6-lts",
+    "@angular/compiler": "v6-lts",
+    "@angular/core": "v6-lts",
+    "@angular/forms": "v6-lts",
+    "@angular/platform-browser": "v6-lts",
+    "@angular/platform-browser-dynamic": "v6-lts",
     "ajv": "^6.9.2",
     "cometd": "^3.1.3",
     "core-js": "^2.5.3",
-    "rxjs": "^5.5.6",
-    "url-polyfill": "^1.0.11",
-    "url-search-params-polyfill": "^2.0.3",
-    "web-animations-js": "^2.3.1",
-    "zone.js": "^0.8.20"
+    "rxjs": "^6.5.5",
+    "url-polyfill": "^1.1.8",
+    "url-search-params-polyfill": "^8.0.0",
+    "web-animations-js": "^2.3.2",
+    "zone.js": "^0.10.3"
   },
   "devDependencies": {
-    "@angular/cli": "^1.7.0",
-    "@angular/compiler-cli": "^5.2.5",
-    "@types/jasmine": "^2.8.6",
+    "@angular-devkit/build-angular": "v6-lts",
+    "@angular/cli": "v6-lts",
+    "@angular/compiler-cli": "v6-lts",
     "@types/node": "^9.4.6",
-    "codelyzer": "^4.1.0",
+    "codelyzer": "latest",
     "font-awesome": "^4.7.0",
-    "jasmine-core": "^2.99.1",
-    "jasmine-spec-reporter": "^4.2.1",
-    "karma": "^2.0.0",
-    "karma-chrome-launcher": "^2.2.0",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.4.1",
-    "karma-jasmine": "~1.1.1",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "^5.3.0",
-    "ts-node": "^5.0.0",
-    "tslint": "^5.9.1",
-    "typescript": "2.5.3"
+    "tslint": "latest",
+    "typescript": "2.9.2"
   }
 }
